app/requirements.txt
python-telegram-bot==20.5
requests
pytz
nest_asyncio